requirements/prod.txt
-r common.txt

wsgiref==0.1.2
gunicorn

django-storages
boto

newrelic
